+++ Domain/Entities/Attendance.cs
@@ -4,7 +4,7 @@
     {
         public DateTimeOffset CheckInTime { get; set; }
         public DateTimeOffset? CheckOutTime { get; set; }
-        public int GuardId { get; set; }
+        public long GuardId { get; set; }
         public Guard Guard { get; set; }
     }
 }
