+++ Domain/Entities/Payroll.cs
@@ -7,7 +7,7 @@
         public decimal HoursWorked { get; set; }
         public decimal TotalPay { get; set; }
         public DateTimeOffset PaymentDate { get; set; }
-        public int GuardId { get; set; }
+        public long GuardId { get; set; }
         public Guard Guard { get; set; }
     }
 }
