+++ Domain/Entities/Report.cs
@@ -6,7 +6,7 @@
         public DateTimeOffset GeneratedDate { get; set; }
         public string Data { get; set; }
         public string Filters { get; set; }
-        public int TenantId { get; set; }
+        public long TenantId { get; set; }
         public Tenant Tenant { get; set; }
     }
 }
